First commit of MVTX alignment calibration
This commit adds a macro to generate MVTX alignment parameters for our simulations.

The MVTX simulation has default alignment ()of perfectly aligned) but this macro generates an XML script with the following PHParameters for any layer and stave position

name = The name of the stave. This defines the ASIC positions, tilts etc
version = The version of the stave GDML file. THis allows us to update indivdual staves if for example, we mask noisy pixels after running for a while
x,y,z position = Thew shift in x, y or z (in cm) relative to perfect placement
tilt = the x-y tilt around the middle of the stave. This may require an x-y shift as well to properly tilt the stave

C. Dean 20210908